rtos/os-decls.h: fix negative constant
- fix clang warnings

+++ include/cmsis-plus/rtos/os-decls.h
@@ -605,7 +605,7 @@ namespace os
        * @details
        * A constant numeric value used to validate the pool size.
        */
-      constexpr size_t max_size = (0 - 1);
+      constexpr size_t max_size = (size_t)(0 - 1);
     } /* namespace mempool */
 
     // ========================================================================
